Fix examples
Fix to define package name.
Add Output test.
